pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "fpdf2>=2.8.5",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "python-nmap>=0.7.1",
    "redis>=7.1.0",
    "reportlab>=4.4.9",
    "scapy>=2.7.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
